Added a jwt flag for loginsuccess to return the token without the need of a redirect.
