+++ lib/controllers/actions/jwt.js
@@ -2,10 +2,6 @@
 
 /* global waterlock */
 
-var moment = require('moment');
-var uuid = require('node-uuid');
-var jwt = require('jwt-simple');
-
 /**
  * jwt action
  * 
@@ -18,22 +14,7 @@ module.exports = function(req, res){
     return res.forbidden('You are not authorized.');
   }
   
-  var jsonWebTokens = waterlock.config.jsonWebTokens || {};
-  var expiryUnit = (jsonWebTokens.expiry && jsonWebTokens.expiry.unit) || 'days';
-  var expiryLength = (jsonWebTokens.expiry && jsonWebTokens.expiry.length) || 7;
-  var expires = moment().add(expiryLength, expiryUnit).valueOf();
-  var issued = Date.now();
-  var user = req.session.user;
-
-  var token = jwt.encode({
-    iss: user.id + '|' + req.remoteAddress,
-    sub: jsonWebTokens.subject, 
-    aud: jsonWebTokens.audience, 
-    exp: expires,
-    nbf: issued,
-    iat: issued,    
-    jti: uuid.v1()
-  }, jsonWebTokens.secret);
+  var token = waterlock._utils.createJwt(req, res);
  
   Jwt.create({token: token, uses: 0, owner: user.id}).exec(function(err){
     if(err){
