+++ lib/cycle.js
@@ -141,7 +141,21 @@ module.exports = function() {
       var postResponse = this._resolvePostAction(waterlock.config.postActions.login.success,
         user);
 
-      if (typeof postResponse === 'string' && this._isURI(postResponse)) {
+      if (postResponse === 'jwt') {
+        //Returns the token immediately
+        var token = waterlock._utils.createJwt(req, res, user);
+
+        Jwt.create({token: token, uses: 0, owner: user.id}).exec(function(err){
+          if(err){
+            return res.serverError('JSON web token could not be created');
+          }
+
+          res.json({
+            token : token,
+            expires: expires
+          });
+        });
+      } else if(typeof postResponse === 'string' && this._isURI(postResponse)){
         res.redirect(postResponse);
       } else {
         res.ok(postResponse);
@@ -312,9 +326,9 @@ module.exports = function() {
      * @return {String|Object} the translated postAction or default value
      * @api private
      */
-    _resolvePostAction: function(mix, def) {
-
-      if (mix === 'default') {
+    _resolvePostAction: function(mix, def){
+      //If postAction is not defined fall back to default
+      if(mix === 'default' || typeof mix === 'undefined') {
         return def;
       }
