+++ lib/utils.js
@@ -1,6 +1,9 @@
 'use strict';
 
 var _ = require('lodash');
+var moment = require('moment');
+var uuid = require('node-uuid');
+var jwt = require('jwt-simple');
 
 /**
  * various utility methods
@@ -68,5 +71,35 @@ module.exports = {
       }
       count++;
     }
+  },
+
+  /**
+   * Creates a new JWT token
+   * 
+   * @param  {Integer} req
+   * @param  {Object} res
+   * @param  {Object} user   the user model
+   * @return {Object}       the created jwt token.
+   * @api public
+   */
+  createJwt: function(req, res, user) {
+    var jsonWebTokens = waterlock.config.jsonWebTokens || {};
+    var expiryUnit = (jsonWebTokens.expiry && jsonWebTokens.expiry.unit) || 'days';
+    var expiryLength = (jsonWebTokens.expiry && jsonWebTokens.expiry.length) || 7;
+    var expires = moment().add(expiryLength, expiryUnit).valueOf();
+    var issued = Date.now();
+    var user = user || req.session.user;
+
+    var token = jwt.encode({
+      iss: user.id + '|' + req.remoteAddress,
+      sub: jsonWebTokens.subject,
+      aud: jsonWebTokens.audience,
+      exp: expires,
+      nbf: issued,
+      iat: issued,
+      jti: uuid.v1()
+    }, jsonWebTokens.secret);
+
+    return token;
   }
 };
