+++ package.json
@@ -10,15 +10,15 @@
       "email": "david.r.rivera193@gmail.com"
     }
   ],
-  "repository":{
+  "repository": {
     "type": "git",
     "url": "https://github.com/waterlock/waterlock"
   },
   "bugs": {
     "url": "https://github.com/waterlock/waterlock/issues"
   },
-  "scripts":{
-    "test":"make test"
+  "scripts": {
+    "test": "make test"
   },
   "bin": {
     "waterlock": "./bin/waterlock.js"
@@ -30,15 +30,15 @@
     "json web tokens",
     "jwt"
   ],
-  "dependencies" : {
-    "moment": "~2.9.0",
-    "commander": "~2.7.1",
-    "lodash": "2.4.1",
+  "dependencies": {
+    "commander": "~2.3.0",
     "jwt-simple": "~0.2.0",
-    "node-uuid": "~1.4.3",
-    "winston": "~0.9.0"
+    "lodash": "2.4.1",
+    "moment": "~2.8.1",
+    "node-uuid": "~1.4.1",
+    "winston": "0.7.3"
   },
-  "devDependencies":{
+  "devDependencies": {
     "mocha": "*",
     "should": "*",
     "proxyquire": "*",
